Changes in Data structure READ Description
Changed many of the int to Integer,
Now it is possible that int will have null value in stead of 0.
ANY time you are checking any property of an individual or family if t's
a string or Integer make sure to check for null before had.
Otherwise you might get a null pointer exception.